HomePage refactorizado
