+++ Frontend/src/components/cards/index.js
@@ -1,2 +1,1 @@
 export * from './ProductCard'
-export * from './SliderCard'
