+++ Frontend/src/components/items/index.js
@@ -1,3 +1,2 @@
 export * from './FeatureItem'
-export * from './StarRating'
 export * from './WebsiteItem'
