--- Frontend/src/components/slider/index.js
@@ -1,1 +1,0 @@
-export * from './SliderHome'
