+++ Frontend/src/layout/index.js
@@ -1,1 +1,2 @@
 export * from './LayoutApp'
+export * from './LayoutPage'
