+++ Frontend/src/pages/Home/assets/index.js
@@ -1,0 +1,1 @@
+export * from './ratingCards'
