+++ Frontend/src/pages/Home/assets/ratingCards.js
@@ -1,0 +1,40 @@
+
+const imageRating = 'src/pages/Home/assets/jesus.jpg'
+
+export const ratingCards = [
+  {
+    image: imageRating,
+    dataValue: 5,
+    title: 'Titulo Card 1',
+    brand: 'Marca',
+    best_price: 5500
+  },
+  {
+    image: imageRating,
+    dataValue: 4.7,
+    title: 'Titulo Card 2',
+    brand: 'Marca',
+    best_price: 7000
+  },
+  {
+    image: imageRating,
+    dataValue: 4.3,
+    title: 'Titulo Card 2',
+    brand: 'Marca',
+    best_price: 5000
+  },
+  {
+    image: imageRating,
+    dataValue: 4,
+    title: 'Titulo Card 4',
+    brand: 'Marca',
+    best_price: 1000
+  },
+  {
+    image: imageRating,
+    dataValue: 3.2,
+    title: 'Titulo Card 5',
+    brand: 'Marca',
+    best_price: 2500
+  }
+]
