+++ Frontend/src/pages/Home/assets/ratingVariant.js
@@ -1,0 +1,2 @@
+export const OFFER_RATING = 'offer'
+export const STAR_RATING = 'star'
