+++ Frontend/src/pages/Home/components/index.js
@@ -1,0 +1,3 @@
+export * from './SliderBox'
+export * from './SliderCard'
+export * from './SliderList'
