+++ Frontend/src/pages/Home/components/rating/StarRating.jsx
@@ -1,14 +1,14 @@
 import { Flex, Icon, useColorModeValue } from '@chakra-ui/react'
 import { MdStar, MdStarBorder, MdStarHalf } from 'react-icons/md'
 
-export const StarRating = ({ rating }) => {
+export const StarRating = ({ value, ...rest }) => {
   const stars = [1, 2, 3, 4, 5]
 
   return (
     <Flex
-      // py={'1px'}
       pt='3px' pb='4px'
       alignItems='center'
+      {...rest}
     >
       {
         stars.map((star, index) => (
@@ -16,21 +16,20 @@ export const StarRating = ({ rating }) => {
             key={index}
             boxSize={4}
             as={
-              (rating >= star)
+              (value >= star)
                 ? MdStar
-                : (rating >= star - 0.5)
+                : (value >= star - 0.5)
                     ? MdStarHalf
                     : MdStarBorder
             }
             color={
-              (rating >= star - 0.5)
+              (value >= star - 0.5)
                 ? useColorModeValue('light.component.active', 'dark.component.active')
                 : useColorModeValue('light.component.main', 'dark.component.main')
-            }
+              }
           />
         ))
       }
-      {/* <Text pl={1} fontSize={14} color={'yellow.500'} fontWeight={'medium'} >{rating}/5</Text> */}
     </Flex>
   )
 }
