+++ Frontend/src/pages/Home/components/rating/index.js
@@ -1,0 +1,2 @@
+export * from './OfferRating'
+export * from './StarRating'
