+++ Frontend/src/pages/Home/components/sliderList.css
