+++ Frontend/src/pages/index.js
@@ -1,4 +1,3 @@
 export * from './Error404Page'
-export * from './HomePage'
 export * from './ProductDetailsPage'
 export * from './ProductsPage'
